building: PKG and COLLECT: refactor check for missing files
Refactor the partially-redundant missing file check in `COLLECT.assemble`;
if the path is referring to an egg-contained resource, the part based on
`not os.path.exists()` will always evaluate to true, rendering the
second part of the check redundant.

This check was modified into its current form by 6a916fc without an
explanation, but the intention was likely to guard against a missing
file error. Therefore, refactor the check to use only `os.path.exists()`
to check for the existence, and display a warning about a missing
resource unless the path is referring to an egg-contained resource.

For the sake of consistency, apply the same refactoring to `PKG.assemble`
that uses a similar (albeit unmodified) check. Also explicitly check
for OPTION typecode instead of checking for empty filename string.