bootloader: delay child signal re-raise until after cleanup
If the child process in the onefile mode is terminated by a signal,
delay re-raising of the signal in the POSIX codepath of the parent
process until after the cleanup. Re-raising it immediately after
the child exits means that clean up is never reached, and the
unpacked temporary directory is left behind by the terminated
frozen application.

Terminating a onefile frozen application by sending SIGTERM to
either the parent or the child process should now clean up
the application's temporary directory. This applies only to the
POSIX codepath, i.e., OSes other than Windows.