building: extend DLL search paths when searching for binary deps
As part of depdendency analysis, we search for binary dependencies
(shared libraries) of the collected binaries, which is done by
running the `find_binary_dependencies` helper function in an
isolated sub-process.

In that helper, we already import all collected top-level packages;
the aim was to ensure that library search paths, which may be
modified during packages' initialization, are up to date. In
practice, however, just importing the packages has little effect
on search paths - at least on Windows, where we still consider
only the binary's parent directory.

Therefore, we now try to explicitly extend the search paths on
Windows, obtaining extra DLL search directories:
 - from the PATH environment variable once all packages are imported
 - by tracking calls to `os.add_dll_directory` calls during package
   imports by monkey-patching that call