bootloader: Windows: implement handling of console ctrl signals
Remove the ingoring of SIGABRT, SIGINT, SIGTERM, and SIGBREAK in
Windows codepath for the parent process of the onefile mode, and
replace it with handling of Windows-specific console control
signals. Ignoring SIGABRT has little value, as it is usually
generated by abnormal conditions within the process (e.g., code
calling `abort()` due to failing validation); so ignoring it is
akin to ignoring bugs in the bootloader code. SIGTERM is, as per
MSDN, never generated on Windows (although it could be raised
programmatically?). And ignoring SIGINT and SIGBREAK prevents us
from properly processing the corresponding console control signals.

For onefile parent process, install the console ctrl signal handler
using the `SetConsoleCtrlHandler`  win32 API function. The installed
handler suppresses CTRL_C_EVENT and CTRL_BREAK_EVENT signals, which
should be equivalent of the old SIGINT and SIGBREAK suppression.
But it also delays the parent process' termination in cases of
CTRL_CLOSE_EVENT (user closing the parent console window),
CTRL_LOGOFF_EVENT (user logging off), and CTRL_SHUTDOWN_EVENT
(system shutting down), in an attempt to buy time for the child
process to exit and for the main thread of the parent process to
perform the clean-up of the temporary directory.