building & hooks: implement module collection mode setting
Implement a mechanism for controlling the collection mode of
modules and packages, with granularity ranging from top-level
packages to individual sub-modules. Therefore, the hooks can
now specify whether the hooked package should be collected as
byte-compiled .pyc modules into embedded PYZ archive (the
default behavior), or as source .py files collected as external
data files (without corresponding modules in the PYZ archive).

The latter option should let us avoid unnecessary .pyc module
collection when the source files are required by the code, or
work around the situations where having a .pyc module in
PYZ archive causes issues due to FrozenImporter's incompatibility
with sys.path manipulation that some packages attempt to perform.

This feature adds a new optional global hook variable, called
`module_collection_mode`. The value can be either a string
("py" or "pyc") or a dictionary of module names and setting
strings.

In the case of a string, the setting affects the hooked module
or a package, and is applied recursively to all sub-packages and
sub-modules, unless another hook overrides it.

The dictionary setting allows a hook to specify different settings
for the package and it subpackages, or even different settings
for other packages.

A corresponding `set_module_collection_mode` method has been
added to the `hook_api` object for adjusting the collection
mode from within the `hook()` function.

The `Analysis` object can now also be passed a dictionary via
an optional `module_collection_mode` argument; the corresponding
settings are applied last, which allows advanced users to both
supplement and override the settings made by the hooks.