building: add support for hash computation in compile_pymodule
If we are compiling py file to pyc, try reading the hash from
the resulting pyc file. If we are given a pyc file, read the hash
from it. If we could not obtain the hash (either the pyc was
timestamp based, or we got code from the code hash), try computing
the hash from the source file.