CArchive: omit PYC header in PYMODULE ('m', 'M') entries
Store PYMODULE ('m', 'M') entries within the CArchive as pure
marshalled code objects instead of full PYC objects (i.e., strip
the PYC header). These entries are used for bootstrap modules and
runtime hooks, and are loaded and executed by the bootloader. There,
we loaded the blob and skipped the PYC header - so there is no
reason to include the PYC header in the first place...