building: clear the source hashes in PYC headers of collected .pyc files
When we collect PYMODULE entries into PYZ, we do not preserve PYC
headers and the source hash information within them. Match this
behavior by zero-ing the source hash in PYC headers when modules
are collected as .pyc files, and always disabling the validation
via the header's flags field.

This puts us at odds with PEP-552 that states "Even when validation
is disabled, the hash field should be set correctly, so out-of-band
consistency checkers can verify the up-to-dateness of the pyc", but
I fail to see any benefit in the case of frozen application.

On the other hand, this saves us some CPU cycles in both
`compile_pymodule` and `create_py3_base_library` helpers, because
we can skip the hash computation when we generate the pyc files from
cached code objects.

This commit effectively reverts an earlier commit that added
support for computing source hashes in .pyc compilation helper.