hooks: add support for setuptools >= 60.0
Add support for setuptools-provided distutils overriding the
deprecated stdlib version, which is the behavior from setuptools
version 60.0 on. As of this version, we need to collect the
vendored setuptools._distutils via hidden imports in order for
setuptools to be importable in frozen application. The installation
of the meta loader that provides override for distutils is handled
in the new runtime hook, pyi_rth_setuptools.py.