tests: fix the scipy test
Remove the `from scipy import *` statement from test_scipy.

Under earlier versions of scipy (< 1.9.0), this import added
nothing in terms of testing the scipy subpackages, as those
were not imported (unless they were explicitly imported by
the top-level package, in which case importing the symbols
again from the top-level package made no difference, either).

With scipy 1.9.0, this causes missing module errors in the
frozen test due to added lazy loading of sub-packages. However,
we have no means of supporting this without always collecting
all scipy subpackages. Plus, this syntax has been and is still
discouraged by the scipy themselves.

If we wanted to thoroughly test scipy subpackages, we would need
to perform a separate test for each subpackage, similarly to as
we do for scikit-learn and scikit-image in the contrib hooks repo.