depend: bytecode: add support for EXTENDED_ARG_QUICK opcode
Python 3.11.0b1 brings EXTENDED_ARG_QUICK specialization of
EXTENDED_ARG, so we need to handle both.

The revised and extended version of this commit also ensures
that all _OPCODES_* variables are used within corresponding
non-capturing groups, even the ones that contain only a single
opcode. In case we ever need to extend the content with multiple
opcodes (as was now the case with _OPCODES_EXTENDED_ARG), the
lack of enclosing group will result in problems with matching.

Also visually reorganize the regexes a bit to make groupings
easier to read; this makes it easier to verify that each opcode
variables use is indeed enclosed with its own non-capturing
group, plus it makes it easier to notice the common patterns.