hookutils: reorganize the Qt hook utilities
Reorganize the Qt module information to provide information necessary
to deal with variations between different python Qt bindings (PySide2,
PyQt5, PySide6, and PyQt6). Replace the existing table-like dictionary
with list of entries, which is easier to format and document. From this
list, we now generate two dictionaries; one that maps Qt module (shared
library) names to the module info entries (the same role as the old
dictionary), and one that maps python module names to the module info
entries. The latter is necessary to accommodate python modules that do
not have corresponding Qt shared libraries (header-only Qt modules,
such as QtAxContainer; or statically-linked module, such as QSci), but
we still need to provide information about plugins or translation
files.

The new information list is based on manual inspection of source code
for Qt 5.15 and 6.3, and should provide comprehensive information about
all plugin names and translation file basenames.

In addition, most of the helper functions, which take a reference to
the `QtLibraryInfo` class as their first argument, have been turned
into methods of the `QtLibraryInfo` class. The corresponding hooks
have also been adjusted.