tests: introduce per-module import tests for Qt bindings
Each module from Qt bindings should be importable on its own; this
requires a corresponding hook that performs recursive analysis of
the module in order to collect all of its dependencies.

The modules from the Qt bindings package are listed dynamically,
which also means that they serve as an indicator whether the
package added a new module during upgrade. In case there is a new
module, it will be without a hook, so its corresponding test
will likely fail due to lack of Qt-library-dependency analysis.