isolated: rework pipe handling
Rework the pipe handling in the `isolated`, primarily in the
Windows-specific codepaths, in an attempt to mitigate the
file handle exhaustion issues that sometimes pop up on the
CI.

Specifically, avoid using `os.pipe()` and instead call WinAPI
`CreatePipe()` via `ctypes`. This gives us file handles without
wrapping them in file descriptors (which are always inheritable
as per python/cpython#77046 and may or may not be related to
our problems).

For good measure, avoid setting `close_fds=False` when spawning
subprocess on Windows; instead, pass the list of relevant file
handles via `subprocess.STARTUPINFO`.

In addition, in both codepaths we now set inheritable flag only
on the file descriptors (or handles) that need to be passed to
the child. Similarly, once the child is spawned, we immediately
close the corresponding file descriptors (or handles) in the
parent process.