depend: rework the module exclusion mechanism
Rework the module exclusion mechanism to immediately suppress excluded
modules during modulegraph's construction (module analysis), rather
as a post-processing modulegraph step. Preventing modules from being
included in the modulegraph in the first place should be more reliable
than trying to remove the nodes and connections from the graph later
on. But most importantly, it ensures that we do not run (potentially
many and potentially costly) hooks for modules that later end up
excluded anyway.

Following suggestion from a FIXME at
https://github.com/pyinstaller/pyinstaller/blob/v5.3/PyInstaller/depend/imphook.py#L497-L519,
this is done by overriding `ModuleGraph._safe_import_hook()` to obtain
the set of excluded imports from the hooks for the referring module and
its parent package(s). The target module's name is then matched against
the excluded imports, where each excluded import entry is treated as a
name prefix. This ensures that the exclusion rules are recursive in both
relevant "directions":
* excluded imports specified in a hook for a package are also applied to
  all submodules and subpackages from that package (and each of those
  may specify additional excluded imports via their own hook)
* if excluded import is a package, we suppress the import of all its
  submodules and subpackages.

The import is suppressed by our `_safe_import_hook()` override returning
an empty list. We cannot return an `ExcludedModule` entry, because that
would preclude proper collection of corresponding module when that module
is available via another different import chain. Returning an empty list
requires a modification of the original modulegraph code, but only in one
place. In other places where `_safe_import_hook()` is called, the source
module is not passed to it, and therefore the modified codepath cannot
be triggered there.

The "basic" modulegraph caching that we perform in order to speed up
our test suite requires that the modulegraph, along with its hooks
cache, is fully serializable via pickle, so that a deep copy can be
performed. Unfortunately, we now load some of our hooks during the
analysis of those basic modules, and the loaded module objects cannot
be pickled.

To work around this, the hook loading is modified to not keep the
references to the loaded modules by default. If loaded due to a magic
attribute look up (i.e., during the modulegraph analysis for excluded
imports), the hook loads the module and copies the attributes
from it. In the post-processing graph step, the hooks are loaded again,
this time keeping the reference to the module, which is necessary for
running the "post-graph API" hook function, if available. In order to
avoid unnecessarily reloading the module that does not have a hook
function, we check for its presence during the first load, and then
reload the hook module during post-graph step only if the hook function
is available.

Consequently, the hooks without hook function are loaded only once, but
hooks with hook function may be loaded twice. However, as the hooks that
provide the hook function should perform heavy lifting within the hook
function, this should have a minimal performance impact.

The post-graph operation step remains, but the code related to removal
of excluded imports has been removed, as it is now unnecessary. Therefore,
post-graph function only updates the magic attributes with values from
the Post-graph API object, stores `binaries` and `datas` to the global
list of files to be collected, and performs analysis of additional modules
specified by the `hiddenimports`. Eventually, the whole post-graph operation
should be merged into base modulegraph's module analysis, but that is
beyond the scope of changes here, and is probably better left for when
try to replace/rewrite the modulegraph itself.