hooks: fix the message in distutils pre-module-find hook
Fix the "retargetting to non-venv dir" message emitted from the
pre-module-find hook in distutils. Due to incorrect comparison
between parent path levels, the message was always emitted, even
when the distutils was found in stdlib directory. Fix that
comparison and handle the case when distutils is provided by
setuptools. Lastly, demote the message from INFO to DEBUG.