hookutils: qt: during analysis, avoid recursing into hooked modules
During recursive analysis of shared libraries that a Qt python
module/extension is linked against, prevent the analysis from
recursing into Qt shared libraries for which we have associated
Qt module information and for which an associated python extension
module exists. In such cases, it should suffice to just add a
hiddenimport; the corresponding hook will perform the extension
module analysis and collect its dependencies.

This prevents repeated analysis of the modules that are linked
against by other modules. For example, under the original
behavior, the hook for QtWebEngineWidgets would end up adding
QtWidgets as a hidden import and analyze QtWidgets (and its
dependencies), thus collecting all plugins from QtWidgets and
translation files from QtCore. Then the hook for QtWidgets would
end up doing the same. Under the new behavior, the hook for
QtWebEngineWidgets only adds QtWidgets to hidden imports, and
lets the hook for the latter perform the remainder of analysis
of that particular branch in dependency tree.