GitHub Workflows security hardening (#7112)
This PR adds explicit permissions section to workflows. This is a security best
practice because by default workflows run with extended set of permissions
(except from on: pull_request from external forks). By specifying any permission
explicitly all others are set to none. By using the principle of least privilege
the damage a compromised workflow can do (because of an injection or compromised
third party tool or action) is restricted. It is recommended to have most strict
permissions on the top level and grant write permissions on job level case by
case.

Signed-off-by: Alex <aleksandrosansan@gmail.com>