tests: add a basic top-level scipy import test
Add a test that imports only top-level scipy package, without any
of the subpackages/modules. This reveals the issue with the new
delvewheel-enabled 1.9.2 Windows wheel, where we do not collect
the external scipy.libs directory (which the delvewheel init code
expects to find).

The existing scipy tests import at least one module that uses an
extension that is in turned linked against the DLLs in that .libs
directory; and due to recent changes in DLL search path handling
and DLL directory structure preservation, we actually end up
collecting (at least some) DLLs into that directory, thus masking
the problem.