hooks: setuptools: do not enable distutils hack for setuptools < 60.0
In the `setuptools` runtime hook, do not attempt to enable the
`distutils` override if `setuptools` version is less than 60.0. In
other words, perform runtime version check, and set the default
value for `SETUPTOOLS_USE_DISTUTILS` based on the major version;
"stdlib" for `setuptools` < 60.0 and "local" for `setuptools` >= 60.0.

This both attempts to mimic the unfrozen behavior (the `distutils`
override was enabled by default in 60.0), but also fixes the error
in frozen application with `setuptools` >= 50.0 and < 60.0, where
our runtime hook would attempt to enable `_distutils_hack` (due to
default value for `SETUPTOOLS_USE_DISTUTILS` being "local",
regardless of the version), but our standard hook forces collection
of `setuptools.distutils˙ only from version 60.0 on.