building: BUNDLE: exempt collected .py/.pyc files from relocation
In macOS .app bundles, exempt the .py and .pyc files that are
collected as DATA files from being relocated from Contents/MacOS
to Contents/Resources (and then symlinked back).

In certain scenarios (such as OpenCV loader's code), the path to
a .py file needs to resolve to the directory that contains the
adjacent binary extensions; i.e., the original directory in
Contents/MacOS, beaucse the extensions are not relocated.

This might introduce regressions with bundle signing and
notarization that we cannot test, hence the breaking change
news entry. But to my knowledge, the main source of problems with
that are dots in the directory names, which should not be the
problem with directory- and base names of the collected .py/.pyc
files.