depend: DLL parent path preservation: work around for pywin32
The `win32uiole` extension is linked agains `win32ui` extension,
but due to the `sys.path` manipulation, the ` win32uiole` is a top
level module, and is collected to top-level directory. Therefore,
the same treatment needs to be applied to the `win32ui` - we should
not preserve its parent directory structure. Apply the same exception
to the whole `site-packages/pythonwin` directory...