bootloader: Windows: fix invalid stream handles in windowed onefile mode
In `onefile` mode, prevent passing invalid stream handles to the
application child process when the standard streams are unavailable
(for example, in a windowed/no-console application).

If streams are invalid/unavailable, the corresponding stream handles
for the child process should be explicitly set to NULL instead of
INVALID_HANDLE_VALUE.

The invalid stream handles cause the `subprocess` module raise an
`OSError: [WinError 6] The handle is invalid` error when trying
to spawn a subprocess in a `onefile` `windowed` frozen application
without redirecting all standard streams. Earlier, we attempted
to mitigate this issue with a runtime hook for `subprocess` that
checks for non-redirected streams and redirects them to devnull;
with the underlying issue fixed, this hook can be removed.