hooks: remove subprocess runtime hook
Remove the `subprocess` runtime hook. The problem with invalid
standard stream handles, which caused the `subprocess` module
raise an `OSError: [WinError 6] The handle is invalid` error
in a `windowed` `onefile` frozen application when trying to
spawn a subprocess without redirecting all standard streams, has
been fixed in the bootloader.

The hook's forced redirection of standard streams was also
interfering with frozen application's attempt to spawn interactive
command prompt, in which case the streams should not have been
redirected.