CI/CD: Automatically lock closed issues after a measure of inactivity. [skip ci]
Almost always, messages on old resolved issues are either off-topic or
people claiming to have "the exact same issue" as some ancient bug
report on an obsolete version of Python but after verbal ping pong
reveal that their problem has nothing in common with the original issue.

Forcing people to go through the process of creating a new issue with
the issue template and filling in their various platform specs should
hopefully dissuade the former and save us time teasing out the details
in the second.