bootloader: implement optional strict unpack mode
Implement a strict unpack mode, where a duplicated file in the
onefile build's PKG triggers a run-time error instead of a warning.

This mode is dynamically enabled at runtime based on the
`PYINSTALLER_STRICT_UNPACK_MODE` environment variable being set
and its value being different from 0.

This feature is primarily intended for the use on our CI as means
for catching file duplication in onefile builds.