building: exclude DEPENDENCY entries from source-file-existence check
For DEPENDENCY entries, we do not need to check whether the source
file exists, as we will not be actually collecting it. Omitting
this check will make it easier to adjust the contents of the TOC
entries and their format (currently, the existence check passes
because the second entry is left unmodified, and thus matches the
original source file).