building: delay merging of reference path and name in DEPENDENCY TOC entry
Within MERGE, do not combine the reference path and target file
name into a single string and store it as the destination name
(the first TOC element). Instead, store the target file name as
destination name (the first TOC element) and the reference path
into the source name (the second TOC element, which is otherwise
left unused for DEPENDENCY TOC entries).

Have the CArchive writer perform the final merge, before writing
the entry to the PKG file.

This ensures that the target name remains unchanged within the
TOC, making it subject of de-duplication codepaths and duplication
checks. Previously, an entry for DEPENDENCY may end up duplicating
another entry (e.g., EXTENSION) at run-time, due to target name
containing the reference path prefix.

We can also get rid of DEPENDENCY-specific handling in `checkCache`
(which returns without any processing if `fnm` contains a colon);
this crutch was needed because `PKG.assemble` incorrectly handled
DEPENDENCY entries and unnecessarily tried running them through
`checkCache`. So we rework that part of `PKG.assemble` to process
DEPENDENCY entries as part of general entry handling. At this point,
this becomes necessary, because even if we kept the hack in
`checkCache`, there is no colon in the `fnm` anymore, so the check
would fail, leading to error...