building: TOC: case-normalize the names for EXTENSION and DEPENDENCY
Turns out that on Windows, the effort of previous commits at
solving the duplication problems is negated by the TOC class
case-normalizing the names only for BINARY and DATA entries,
but not for EXTENSION and DEPENDENCY ones.

Fix that by adding EXTENSION and DEPENDENCY to the list of typecodes
whose names need to be case-normnalized before comparison against
existing names.