isolated: introduce time-out in the wait for the child subprocess to exit
Prevent isolated-subprocess calls from indefinitely blocking in their
clean-up codepath when the subprocess fails to exit. After the grace
period of 5 seconds, we now attempt to terminate such subprocess in
order to prevent hanging of the build process.

For the purposes of our CI, introduce a strict mode (by default
controlled by the ˙compat.strict_collect_mode˙ and its backing
environment variable), which raises an error instead of just
printing the warning.