building: cleanup remove pyc argument from _check_guts_toc_mtime
The only place where we use `_check_guts_toc_mtime` with `pyc`
argument enabled is when checking the `Analysis.pure` TOC, and
the source names of those entries already point to source .py files.
So shortening the filenames by one character results in checking
for non-existant .p files.

Even if an entry happened to point to a .pyc file, it is highly
unlikely that there would be an adjacent .py file available,
because under contemporary python 3 versions, that would hide the
.pyc file from the loader.