building: MERGE: remove common path computation
The MERGE code seems to go out of its way to determine the common
path prefix of the entry-point scripts from the given Analysis
objects. And then for each passed Analysis object, use this prefix
to attempt to compute the path to the executable. Or, if this
computed path matches the passed identifier (2nd tuple element),
uses it to look up the passed path-to-executable argument (3rd
tuple element).

However, this all seems to be unnecessary, because the entry-point
script name does not necessarily match the executable's name, so
we need to pass the path-to-executable anyway, and these paths
are actually what determines proper relative paths for shared
dependency references. So in cases when the script-based path
matches the passed identifier, we end up using the passed
path-to-executable, and when it does not, the used script-based
path very likely results in incorrect references.

So remove all the common-path-related code, and rely only on
passed paths-to-executables.