building: PYZ/EXE/PKG/Splash: remove the .typ class member
Remove the .typ class member from PYZ, EXE, PKG, and Splash.
These were used to dynamically identify sub-type of the Target
when checking the passed arguments.

However, the Target class itself does not have this member,
which means that every location where it is used, we are
already implicitly expecting only certain sub-classes of
Target. So it is more straightforward to directly use
isinstance() checks to identifiy the potential classes of
interest, with the added benefit of being able to raise
incompatible-argument errors for unhandled Target instances.