building: EXE: move check for unsupported properties to constructor
Move the check for platform-unsupported properties (i.e., version
info, manifest, and resources are supported only on Windows; icon
is supported only on Windows and macOS) from guts change check to
the  constructor.

This way, we always clear the unsupported properties, and we also
emit the warning on every build instead of only on repeated builds
during cached state comparison.