building: EXE: load version info structure before comparing guts
Load the version information structure in `EXE` constructor, so that
the comparison against cached state is done with the structure instead
of the filen name. This way, changing the contents of the version
information file triggers rebuild of the EXE.

Split and clean-up related functions in the `utils.win32.versioninfo`
module as well as in `pyi-grab_version` and `pyi-set_version`
utility scripts.