tests: remove test_threading_import
The test was testing the multi-threaded read behavior of the
ArchiveFile, which was used by archive readers to dynamically
open and close the archive file when reading from it. The
ArchiveFile wrapper is gone now (it is easier and less error-prone
to just open the file directly whenever needed), so remove the
test as well.