bootloader: wscript: make cmocka-based tests opt-in
Build the bootloader's cmocka-based test suite only if the tests
are explicitly enabled via the --tests command-line switch (and
add corresponding --no-tests switch for the sake of completeness).

This prevents the test suite from being automatically built when
the cmocka library is found, which is undesirable is some
situations; e.g., on macOS, homebrew-installed cmocka is
single-arch, which results in errors when trying to build bootloader
and its test suite in universal2 mode.