hookutils: tcl/tk: fix find_tcl_tk_shared_libs for cygwin
When trying to resolve link-time dependencies of _tkinter extension,
`bindepend.getfullnameof()` needs to be used only for Windows. On
other systems, including Cygwin, we already obtain the full path
to the library.

Furthermore, `bindepend.getfullnameof()` requires `pywin32-ctypes`
(due to call to `compat.win32api.GetSystemDirectory()`), which is
not available under cygwin.