Prevent pyinstaller-hooks-contrib *randomly* shadowing hooks from packages.
When both pyinstaller-hooks-contrib and a library contain a hook for
that library, the hook from the library should take priority. That way a
package which has a hook in pyinstaller-hooks-contrib can switch to
taking ownership of their hook without our (likely out of date) version
clobbering their efforts.

Currently, they have the same priority meaning that which hook is
actually used is non deterministic.