hooks: rework the multiprocessing run-time hook
Provide custom override of the `Popen` only for `spawn` and
`forkserver` start methods. These run the `sys.executable` from
scratch, and therefore need the `_MEIPASS2` environment variable
restored.

In contrast, we do not need to override `Popen` for `fork`, which
forks the current process, without re-running the executable.
Even more, forking the process means that the child process
inherits a copy of the lock in the `FrozenSupportMixIn` class;
for nested multiprocess calls using the `fork` method, the lock
is copied in the locked state, leading to deadlock.

At the same time, reorganize the hook by wrapping all its
functionality into a function that is executed and deleted
afterwards. This minimizes the pollution of the global namespace
with elements from the run-time hook.