tests: multiprocessing: add test for retrieving __main__' code object
Add a test that retrieves (and disassembles) the code object of
the `__main__` module within the `multiprocessing.Process`. This
is similar to the existing `test_misc.py::test_dis_main`, except
it tests for availability of `__main__`'s code object within the
process.

When ran unfrozen, the test script seems to work only with the
`fork` start method. With current implementation of `FrozenImporter`
and `multiprocessing` support, however, it seems to work with all
start methods when frozen. Remove or adjust it accordingly if we
ever change the behavior.