hooks: update PySide6 hooks for PySide6 6.5.0
Add hooks for `PySide6.QtLocation` and `PySide6.QtSerialBus` that
are available in 6.5.0, and update corresponding module definitions.

Also mark `QtPositioning` python module as available on all bindings,
since that seems to be the case (and has been for a while?).

As a side note, the preview of `PyQt6` 6.5.0 does not seem to
provide `QtLocation` nor `QtSerialBus` python modules (nor the
corresponding Qt shared libraries), so no change need to be made
for `PyQt6` yet.