bootloader: linux: ignore /proc/self/exe when launched via ld.so
When PyInstaller-frozen executable is launched via ld.so dynamic
loader (e.g., `/lib64/ld-linux-x86-64.so.2 ./frozen_program`),
the result of `/proc/self/exe` name resolution needs to be
ignored, as it points to the `ld.so` program instead of the
actual PyInstaller executable.

Therefore, we now try to detect if `/proc/self/exe` name matches
the `ld.so` pattern (`ld-*.so.*`), and if it does, fall back to
regular name analysis. In this fall-back mode, if the executable
is a symbolic link, we need to explicitly resolve it ourselves.