tests: Qt: always initialize QApplication with sys.argv
Looks like passing an empty list is causing problems with QtWebEngine
in PyQt6 6.5.0. And even if we can get away with it in other tests
and with other bindings, it makes more sense to properly pass
`sys.argv`, as one would in a real application. Applies to all
instantiations of `QCoreApplication`, `QGuiApplication`, and
`QApplication`.